Remove builders and engines logic
